Move client creation to crate root
